# LUCENE-10297: Speed up medium cardinality fields with readLongs and SIMD