--- a/lucene/core/src/java/org/apache/lucene/util/IntCodecHelper.java
+++ b/lucene/core/src/java/org/apache/lucene/util/IntCodecHelper.java
@@ -1,9 +1,8 @@
 package org.apache.lucene.util;
 
+import java.io.IOException;
 import org.apache.lucene.store.DataOutput;
 import org.apache.lucene.store.IndexInput;
-
-import java.io.IOException;
 
 public class IntCodecHelper {
 
@@ -11,20 +10,20 @@ public class IntCodecHelper {
   private static final int BLOCK_SIZE_MINUS_ONE = BLOCK_SIZE - 1;
 
   public static void writeInts8(DataOutput out, int count, int[] values, int offset)
-          throws IOException {
+      throws IOException {
     int k = 0;
     for (; k < count - BLOCK_SIZE_MINUS_ONE; k += BLOCK_SIZE) {
       int step = offset + k;
       for (int i = 0; i < 16; ++i) {
         long l =
-                ((long) values[step + i] << 56)
-                        | ((long) values[step + 16 + i] << 48)
-                        | ((long) values[step + 32 + i] << 40)
-                        | ((long) values[step + 48 + i] << 32)
-                        | ((long) values[step + 64 + i] << 24)
-                        | ((long) values[step + 80 + i] << 16)
-                        | ((long) values[step + 96 + i] << 8)
-                        | (long) values[step + 112 + i];
+            ((long) values[step + i] << 56)
+                | ((long) values[step + 16 + i] << 48)
+                | ((long) values[step + 32 + i] << 40)
+                | ((long) values[step + 48 + i] << 32)
+                | ((long) values[step + 64 + i] << 24)
+                | ((long) values[step + 80 + i] << 16)
+                | ((long) values[step + 96 + i] << 8)
+                | (long) values[step + 112 + i];
         out.writeLong(l);
       }
     }
@@ -34,16 +33,16 @@ public class IntCodecHelper {
   }
 
   public static void writeInts16(DataOutput out, int count, int[] values, int offset)
-          throws IOException {
+      throws IOException {
     int k = 0;
     for (; k < count - BLOCK_SIZE_MINUS_ONE; k += BLOCK_SIZE) {
       int step = offset + k;
       for (int i = 0; i < 32; ++i) {
         long l =
-                ((long) values[step + i] << 48)
-                        | ((long) values[step + 32 + i] << 32)
-                        | ((long) values[step + 64 + i] << 16)
-                        | (long) values[step + 96 + i];
+            ((long) values[step + i] << 48)
+                | ((long) values[step + 32 + i] << 32)
+                | ((long) values[step + 64 + i] << 16)
+                | (long) values[step + 96 + i];
         out.writeLong(l);
       }
     }
@@ -53,7 +52,7 @@ public class IntCodecHelper {
   }
 
   public static void writeInts32(DataOutput out, int count, int[] values, int offset)
-          throws IOException {
+      throws IOException {
     int k = 0;
     for (; k < count - BLOCK_SIZE_MINUS_ONE; k += BLOCK_SIZE) {
       int step = offset + k;
@@ -68,7 +67,7 @@ public class IntCodecHelper {
   }
 
   public static void readInts8(IndexInput in, int count, long[] values, int offset)
-          throws IOException {
+      throws IOException {
     int k = 0;
     for (; k < count - BLOCK_SIZE_MINUS_ONE; k += BLOCK_SIZE) {
       final int step = offset + k;
@@ -91,7 +90,7 @@ public class IntCodecHelper {
   }
 
   public static void readInts16(IndexInput in, int count, long[] values, int offset)
-          throws IOException {
+      throws IOException {
     int k = 0;
     for (; k < count - BLOCK_SIZE_MINUS_ONE; k += BLOCK_SIZE) {
       int step = offset + k;
@@ -110,7 +109,7 @@ public class IntCodecHelper {
   }
 
   public static void readInts32(IndexInput in, int count, long[] values, int offset)
-          throws IOException {
+      throws IOException {
     int k = 0;
     for (; k < count - BLOCK_SIZE_MINUS_ONE; k += BLOCK_SIZE) {
       final int step = offset + k;