… lucene/CHANGES.txt
… lucene/core/src/java/org/apache/lucene/util/bkd/Run.java
… lucene/core/src/test/org/apache/lucene/util/bkd/TestDocIdsWriter.java